bug identified and new workflow to run sex using customized config files
There is a known issue. The crossmatch_diff_table creates columns objectHCG_1_1 and objectHCG_1_2.  Then the artificat Select_columns artificat selects only the objectHCG_1_1 column to perform the following crossmatches. This is a problem for those cases where objectHCG_1_1 has no value. This may happen if the check_results_comparing_Band_r_i workflow doesn't agree with check_results_comparing_band_r_g workflow and a galaxy is only tagged as bad row in one case. In particular, the problem appears if a galaxy is only identified as bad row in the r_g comparition.  Then, it will pass as a good result while it should be given as bad row

	new file:   datasets/inputs/gather_galaxy_properties/tableforCustomizedListOfConfigFiles.xml
	new file:   workflows/Gather_HCG_galaxy_properties_using_existing_sexConfigFile.t2flow